Removing unused Ast class, and moving concrete implementation of apply() method into AstBuiltin class
