+++ h2o-core/src/main/java/water/rapids/ast/AstPrimitive.java
@@ -1,6 +1,5 @@
 package water.rapids.ast;
 
-import water.H2O;
 import water.rapids.Env;
 import water.rapids.Val;
 import water.rapids.vals.ValFun;
@@ -32,15 +31,6 @@ public abstract class AstPrimitive<T extends AstPrimitive<T>> extends AstRoot<T>
    * <p>Primary method to invoke this function, passing all the parameters
    * as the `asts` list.</p>
    *
-   * <p>The default implementation of this method executes all Asts within
-   * the provided environment, and then calls {@link #apply(Val[])} passing it
-   * the arguments as the list of {@link Val}s. A derived class will then only
-   * need to override the second `apply()` function which is much simpler.</p>
-   *
-   * <p>However for certain functions (such as short-circuit boolean operators)
-   * executing all arguments is not desirable -- these functions would have to
-   * override this more general method.</p>
-   *
    * @param env Current execution environment. Variables are looked up here.
    * @param stk TODO need clarification
    * @param asts List of AstRoot expressions that are arguments to the
@@ -48,24 +38,7 @@ public abstract class AstPrimitive<T extends AstPrimitive<T>> extends AstRoot<T>
    * @return value resulting from calling the function with the provided list
    *         of arguments.
    */
-  public Val apply(Env env, Env.StackHelp stk, AstRoot[] asts) {
-    Val[] args = new Val[asts.length];
-    args[0] = null;
-    for (int i = 1; i < asts.length; i++) {
-      args[i] = stk.track(asts[i].exec(env));
-    }
-    return apply(args);
-  }
-
-  /**
-   * Most Ast* functions will want to override this method. The semantics is
-   * "call this function with the provided list of arguments".
-   */
-  @SuppressWarnings("UnusedParameters")
-  protected Val apply(Val[] args) {
-    throw H2O.unimpl();
-  }
-
+  public abstract Val apply(Env env, Env.StackHelp stk, AstRoot[] asts);
 
 
   @Override
